+++ packages/ui/src/alert-dialog/trigger.tsx
@@ -4,10 +4,10 @@ import * as Dialog from '../dialog';
 import type { TriggerProps } from './types';
 
 /**
- * Renders a button that opens the confirmation dialog when clicked.
+ * Renders a button that opens the alert dialog when clicked.
  */
 const Trigger = forwardRef< HTMLButtonElement, TriggerProps >(
-	function ConfirmDialogTrigger( props, ref ) {
+	function AlertDialogTrigger( props, ref ) {
 		return <Dialog.Trigger ref={ ref } { ...props } />;
 	}
 );