+++ packages/ui/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### New Features
 
--   Add `ConfirmDialog` primitive ([#76847](https://github.com/WordPress/gutenberg/pull/76847)).
+-   Add `AlertDialog` primitive ([#76847](https://github.com/WordPress/gutenberg/pull/76847)).
 -   Add `InputControl` component ([#76653](https://github.com/WordPress/gutenberg/pull/76653)).
 -   `Dialog`: Expose `initialFocus` and `finalFocus` props on `Dialog.Popup` for custom focus management ([#76860](https://github.com/WordPress/gutenberg/pull/76860)).